Implement RegisterProcessingService (#117)
* feat: add register processing service

* wip: fix tests
