+++ Logibooks.Core.Tests/Controllers/RegistersControllerTests.cs
@@ -15,8 +15,7 @@
 // 'AS IS' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
 // TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
 // PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS
-// BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
 // SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
 // INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
 // CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
@@ -41,7 +40,6 @@ using System.IO;
 using System.Threading;
 using System;
 using System.Linq;
-using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Logibooks.Core.Tests.Controllers;
@@ -54,17 +52,18 @@ public class RegistersControllerTests
     private Mock<IHttpContextAccessor> _mockHttpContextAccessor;
     private Mock<IRegisterValidationService> _mockRegValidationService;
     private ILogger<RegistersController> _logger;
-    private RegistersController _controller;
     private Role _logistRole;
     private Role _adminRole;
     private User _logistUser;
     private User _adminUser;
-    private Type _controllerType;
-    private MethodInfo _processExcelMethod;
+    private RegistersController _controller;
+    private Mock<IRegisterProcessingService> _mockProcessingService;
 #pragma warning restore CS8618
 
     private readonly string testDataDir = Path.Combine(AppContext.BaseDirectory, "test.data");
 
+
+    // Initialize the _controller field in the Setup method
     [SetUp]
     public void Setup()
     {
@@ -85,7 +84,7 @@ public class RegistersControllerTests
             Password = hpw,
             FirstName = "Log",
             LastName = "User",
-            UserRoles = [ new UserRole { UserId = 1, RoleId = 1, Role = _logistRole } ]
+            UserRoles = [new UserRole { UserId = 1, RoleId = 1, Role = _logistRole }]
         };
         _adminUser = new User
         {
@@ -94,21 +93,16 @@ public class RegistersControllerTests
             Password = hpw,
             FirstName = "Adm",
             LastName = "User",
-            UserRoles = [ new UserRole { UserId = 2, RoleId = 2, Role = _adminRole } ]
+            UserRoles = [new UserRole { UserId = 2, RoleId = 2, Role = _adminRole }]
         };
         _dbContext.Users.AddRange(_logistUser, _adminUser);
         _dbContext.SaveChanges();
 
         _mockHttpContextAccessor = new Mock<IHttpContextAccessor>();
         _mockRegValidationService = new Mock<IRegisterValidationService>();
+        _mockProcessingService = new Mock<IRegisterProcessingService>();
         _logger = new LoggerFactory().CreateLogger<RegistersController>();
-        _controller = new RegistersController(_mockHttpContextAccessor.Object, _dbContext, _logger, _mockRegValidationService.Object);
-
-        _controllerType = typeof(RegistersController);
-        _processExcelMethod = _controllerType.GetMethod("ProcessExcel",
-            BindingFlags.NonPublic | BindingFlags.Instance)
-            ?? throw new InvalidOperationException("ProcessExcel method not found");
-
+        _controller = new RegistersController(_mockHttpContextAccessor.Object, _dbContext, _logger, _mockRegValidationService.Object, _mockProcessingService.Object);
     }
 
     [TearDown]
@@ -123,7 +117,7 @@ public class RegistersControllerTests
         var ctx = new DefaultHttpContext();
         ctx.Items["UserId"] = id;
         _mockHttpContextAccessor.Setup(x => x.HttpContext).Returns(ctx);
-        _controller = new RegistersController(_mockHttpContextAccessor.Object, _dbContext, _logger, _mockRegValidationService.Object);
+        _controller = new RegistersController(_mockHttpContextAccessor.Object, _dbContext, _logger, _mockRegValidationService.Object, _mockProcessingService.Object);
     }
 
     [Test]
@@ -695,13 +689,25 @@ public class RegistersControllerTests
             return;
         }
 
+        // Set up the mock processing service to return a successful reference
+        var expectedReference = new Reference { Id = 123 };
+        _mockProcessingService.Setup(x => x.UploadWbrRegisterFromExcelAsync(
+            It.IsAny<int>(), 
+            It.IsAny<byte[]>(), 
+            It.IsAny<string>(), 
+            It.IsAny<string>(), 
+            It.IsAny<CancellationToken>()))
+            .ReturnsAsync(expectedReference);
+
         var mockFile = CreateMockFile("Реестр_207730349.xlsx", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", excelContent);
         var result = await _controller.UploadRegister(mockFile.Object);
         Assert.That(result, Is.TypeOf<CreatedAtActionResult>());
 
-        // Verify that orders were created in the database
-        Assert.That(_dbContext.Orders.Count(), Is.GreaterThan(0),
-            "Orders should have been created in the database");
+        // Verify the CreatedAtActionResult properties
+        var createdResult = result as CreatedAtActionResult;
+        Assert.That(createdResult!.Value, Is.TypeOf<Reference>());
+        var returnedReference = createdResult.Value as Reference;
+        Assert.That(returnedReference!.Id, Is.EqualTo(expectedReference.Id));
     }
 
     [Test]
@@ -750,6 +756,16 @@ public class RegistersControllerTests
             return;
         }
 
+        // Set up the mock processing service to return a successful reference
+        var expectedReference = new Reference { Id = 124 };
+        _mockProcessingService.Setup(x => x.UploadWbrRegisterFromExcelAsync(
+            It.IsAny<int>(), 
+            It.IsAny<byte[]>(), 
+            It.IsAny<string>(), 
+            It.IsAny<string>(), 
+            It.IsAny<CancellationToken>()))
+            .ReturnsAsync(expectedReference);
+
         var mockFile = CreateMockFile("Реестр_207730349.zip", "application/zip", zipContent);
 
         var result = await _controller.UploadRegister(mockFile.Object);
@@ -757,8 +773,11 @@ public class RegistersControllerTests
         // Assert that the result is OK
         Assert.That(result, Is.TypeOf<CreatedAtActionResult>());
 
-
-        Assert.That(_dbContext.Orders.Count(), Is.GreaterThan(0));
+        // Verify the CreatedAtActionResult properties
+        var createdResult = result as CreatedAtActionResult;
+        Assert.That(createdResult!.Value, Is.TypeOf<Reference>());
+        var returnedReference = createdResult.Value as Reference;
+        Assert.That(returnedReference!.Id, Is.EqualTo(expectedReference.Id));
     }
 
     [Test]
@@ -778,6 +797,15 @@ public class RegistersControllerTests
             Assert.Fail($"Test file not found at {testFilePath}: {ex.Message}");
             return;
         }
+
+        // Set up the mock processing service to throw InvalidOperationException for empty files
+        _mockProcessingService.Setup(x => x.UploadWbrRegisterFromExcelAsync(
+            It.IsAny<int>(), 
+            It.IsAny<byte[]>(), 
+            It.IsAny<string>(), 
+            It.IsAny<string>(), 
+            It.IsAny<CancellationToken>()))
+            .ThrowsAsync(new InvalidOperationException("Excel file is empty"));
 
         var mockFile = CreateMockFile("Register_Empty.xlsx", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", excelContent);
         var result = await _controller.UploadRegister(mockFile.Object);
@@ -877,45 +905,6 @@ public class RegistersControllerTests
         Assert.That(registersCount, Is.EqualTo(0), "No register should be created for unsupported file types");
     }
 
-    private async Task<IActionResult> InvokeProcessExcel(int companyId, byte[] content, string fileName, string mappingFile = "wbr_register_mapping.yaml")
-    {
-        return await (Task<IActionResult>)_processExcelMethod.Invoke(
-            _controller,
-            [companyId, content, fileName, mappingFile])!;
-    }
-
-    [Test]
-    public async Task ProcessExcel_Returns500Error_WhenMappingFileNotFound()
-    {
-        // Arrange
-        // Create a sample Excel file content
-        byte[] excelContent = [0x50, 0x4B, 0x03, 0x04]; // Just some dummy content
-
-        // Make sure mapping directory exists but use a non-existent mapping file name
-        string mappingDir = Path.Combine(AppContext.BaseDirectory, "mapping");
-        Directory.CreateDirectory(mappingDir);
-        string nonExistentMappingFile = "non_existent_mapping.yaml";
-        string mappingPath = Path.Combine(mappingDir, nonExistentMappingFile);
-
-        // Delete the mapping file if it somehow exists
-        if (File.Exists(mappingPath))
-        {
-            File.Delete(mappingPath);
-        }
-
-        // Act
-        var result = await InvokeProcessExcel(1, excelContent, "test.xlsx", nonExistentMappingFile);
-
-        // Assert
-        Assert.That(result, Is.TypeOf<ObjectResult>());
-        var objResult = result as ObjectResult;
-        Assert.That(objResult!.StatusCode, Is.EqualTo(StatusCodes.Status500InternalServerError));
-
-        var errMessage = objResult.Value as ErrMessage;
-        Assert.That(errMessage, Is.Not.Null);
-        Assert.That(errMessage!.Msg, Does.Contain("Не найдена спецификация файла реестра"));
-        Assert.That(errMessage.Msg, Does.Contain(mappingPath));
-    }
 
     [Test]
     public async Task DeleteRegister_DeletesEmptyRegister_WhenUserIsLogist()
@@ -1182,7 +1171,7 @@ public class RegistersControllerTests
 
         // Update the logger type to match the expected type for RegisterValidationService
         var realRegSvc = new RegisterValidationService(_dbContext, scopeFactoryMock.Object, new LoggerFactory().CreateLogger<RegisterValidationService>(), new MorphologySearchService(), new FeacnPrefixCheckService(_dbContext));
-        _controller = new RegistersController(_mockHttpContextAccessor.Object, _dbContext, _logger, realRegSvc);
+        _controller = new RegistersController(_mockHttpContextAccessor.Object, _dbContext, _logger, realRegSvc, _mockProcessingService.Object);
 
         var result = await _controller.ValidateRegister(200);
         var handle = ((GuidReference)((OkObjectResult)result.Result!).Value!).Id;
@@ -1203,118 +1192,3 @@ public class RegistersControllerTests
     }
 }
 
-[TestFixture]
-public class ConvertValueToPropertyTypeTests
-{
-#pragma warning disable CS8618
-    private RegistersController _controller;
-#pragma warning restore CS8618
-
-    [SetUp]
-    public void Setup()
-    {
-        // Create a minimal DbContextOptions for AppDbContext
-        var options = new DbContextOptionsBuilder<AppDbContext>()
-            .UseInMemoryDatabase($"test_db_{Guid.NewGuid()}")
-            .Options;
-
-        // Pass the valid options to the AppDbContext constructor
-        var dbContext = new AppDbContext(options);
-
-        // Create the controller instance with the valid AppDbContext
-        _controller = (RegistersController)Activator.CreateInstance(
-            typeof(RegistersController),
-            [new Mock<IHttpContextAccessor>().Object, dbContext, new Mock<ILogger<RegistersController>>().Object, new Mock<IRegisterValidationService>().Object]
-        )!;
-    }
-
-    [TestCase("42", typeof(int), 42)]
-    [TestCase("notanint", typeof(int), 0)]
-    [TestCase("3.14", typeof(double), 3.14)]
-    [TestCase("2,71", typeof(double), 2.71)] // comma as decimal separator
-    [TestCase("notadouble", typeof(double), 0.0)]
-    [TestCase("123.45", typeof(decimal), 123.45)]
-    [TestCase("67,89", typeof(decimal), 67.89)]
-    [TestCase("notadecimal", typeof(decimal), 0.0)]
-    [TestCase("true", typeof(bool), true)]
-    [TestCase("false", typeof(bool), false)]
-    [TestCase("1", typeof(bool), true)]
-    [TestCase("0", typeof(bool), false)]
-    [TestCase("yeS", typeof(bool), true)]
-    [TestCase("no", typeof(bool), false)]
-    [TestCase("Да", typeof(bool), true)]
-    [TestCase("нет", typeof(bool), false)]
-    [TestCase("", typeof(bool), false)]
-    [TestCase("notabool", typeof(bool), false)]
-    [TestCase("2024-06-28", typeof(DateTime), "2024-06-28")]
-    [TestCase("notadate", typeof(DateTime), "0001-01-01")]
-    [TestCase("2024-06-28", typeof(DateOnly), "2024-06-28")]
-    [TestCase("2024-06-28T13:00:12", typeof(DateOnly), "2024-06-28")]
-    [TestCase("notadate", typeof(DateOnly), "0001-01-01")]
-    [TestCase("hello", typeof(string), "hello")]
-    [TestCase("", typeof(string), "")]
-    [TestCase(null, typeof(string), "")]
-
-    public void ConvertValueToPropertyType_PrimitiveTypes_Works(string? input, Type type, object expected)
-    {
-        var result = _controller.GetType()
-            .GetMethod("ConvertValueToPropertyType", BindingFlags.NonPublic | BindingFlags.Instance)!
-            .Invoke(_controller, new object?[] { input, type, "TestProp" });
-
-        if (type == typeof(DateTime))
-        {
-            var expectedDate = DateTime.TryParse(expected.ToString(), out var dt) ? dt : default;
-            Assert.That(result, Is.EqualTo(expectedDate));
-        }
-        else if (type == typeof(DateOnly))
-        {
-            var expectedDate = DateOnly.TryParse(expected.ToString(), out var d) ? d : default;
-            Assert.That(result, Is.EqualTo(expectedDate));
-        }
-        else
-        {
-            Assert.That(result, Is.EqualTo(expected));
-        }
-    }
-
-    [Test]
-    public void ConvertValueToPropertyType_NullableInt_ReturnsNullOnNull()
-    {
-        var type = typeof(int?);
-        var result = _controller.GetType()
-            .GetMethod("ConvertValueToPropertyType", BindingFlags.NonPublic | BindingFlags.Instance)!
-            .Invoke(_controller, new object?[] { null, type, "TestProp" });
-        Assert.That(result, Is.Null);
-    }
-
-    [Test]
-    public void ConvertValueToPropertyType_NullableDouble_ReturnsNullOnEmpty()
-    {
-        var type = typeof(double?);
-        var result = _controller.GetType()
-            .GetMethod("ConvertValueToPropertyType", BindingFlags.NonPublic | BindingFlags.Instance)!
-            .Invoke(_controller, new object?[] { "", type, "TestProp" });
-        Assert.That(result, Is.Null);
-    }
-
-    [Test]
-    public void ConvertValueToPropertyType_UnknownType_UsesChangeType()
-    {
-        var type = typeof(long);
-        var result = _controller.GetType()
-            .GetMethod("ConvertValueToPropertyType", BindingFlags.NonPublic | BindingFlags.Instance)!
-            .Invoke(_controller, new object?[] { "123456789", type, "TestProp" });
-        Assert.That(result, Is.EqualTo(123456789L));
-    }
-
-    [Test]
-    public void ConvertValueToPropertyType_UnknownType_ReturnsDefaultOnError()
-    {
-        var type = typeof(Guid);
-        var result = _controller.GetType()
-            .GetMethod("ConvertValueToPropertyType", BindingFlags.NonPublic | BindingFlags.Instance)!
-            .Invoke(_controller, new object?[] { "notaguid", type, "TestProp" });
-        Assert.That(result, Is.EqualTo(Guid.Empty));
-    }
-
-}
