+++ Logibooks.Core/Controllers/RegistersController.cs
@@ -25,16 +25,12 @@
 
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using System.Globalization;
-
 using SharpCompress.Archives;
-using ExcelDataReader;
 
 using Logibooks.Core.Authorization;
 using Logibooks.Core.Data;
 using Logibooks.Core.Models;
 using Logibooks.Core.RestModels;
-using Logibooks.Core.Settings;
 using Logibooks.Core.Services;
 
 namespace Logibooks.Core.Controllers;
@@ -48,13 +44,15 @@ public class RegistersController(
     IHttpContextAccessor httpContextAccessor,
     AppDbContext db,
     ILogger<RegistersController> logger,
-    IRegisterValidationService validationService) : LogibooksControllerBase(httpContextAccessor, db, logger)
+    IRegisterValidationService validationService,
+    IRegisterProcessingService processingService) : LogibooksControllerBase(httpContextAccessor, db, logger)
 {
 
     private readonly string[] allowedSortBy = ["id", "filename", "date", "orderstotal"];
     private readonly int maxPageSize = 100;
     private readonly int idWBR = 2;
     private readonly IRegisterValidationService _validationService = validationService;
+    private readonly IRegisterProcessingService _processingService = processingService;
 
     [HttpGet("{id}")]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(RegisterViewItem))]
@@ -250,8 +248,21 @@ public class RegistersController(
             using var ms = new MemoryStream();
             await file.CopyToAsync(ms);
             byte[] excelContent = ms.ToArray();
-            var result = await ProcessExcel(companyId, excelContent, file.FileName);
-            return result;
+            try
+            {
+                var reference = await _processingService.UploadWbrRegisterFromExcelAsync(companyId, excelContent, file.FileName);
+                return CreatedAtAction(nameof(UploadRegister), new { id = reference.Id }, reference);
+            }
+            catch (FileNotFoundException ex)
+            {
+                _logger.LogError("Mapping file not found at {path}, UploadRegister returning '500 Internal Server Error'", ex.FileName);
+                return _500Mapping(ex.FileName ?? string.Empty);
+            }
+            catch (InvalidOperationException)
+            {
+                _logger.LogDebug("UploadRegister returning '400 Bad Request' - Excel file is empty");
+                return _400EmptyRegister();
+            }
         }
         else if (fileExtension == ".zip" || fileExtension == ".rar")
         {
@@ -285,9 +296,22 @@ public class RegistersController(
                 excelContent = entryStream.ToArray();
             }
 
-            var result = await ProcessExcel(companyId, excelContent, excelFileName);
-            _logger.LogDebug("UploadRegister processed archive with Excel");
-            return result;
+            try
+            {
+                var reference = await _processingService.UploadWbrRegisterFromExcelAsync(companyId, excelContent, excelFileName);
+                _logger.LogDebug("UploadRegister processed archive with Excel");
+                return CreatedAtAction(nameof(UploadRegister), new { id = reference.Id }, reference);
+            }
+            catch (FileNotFoundException ex)
+            {
+                _logger.LogError("Mapping file not found at {path}, UploadRegister returning '500 Internal Server Error'", ex.FileName);
+                return _500Mapping(ex.FileName ?? string.Empty);
+            }
+            catch (InvalidOperationException)
+            {
+                _logger.LogDebug("UploadRegister returning '400 Bad Request' - Excel file is empty");
+                return _400EmptyRegister();
+            }
         }
         else
         {
@@ -455,165 +479,4 @@ public class RegistersController(
                 g => g.ToDictionary(x => x.StatusId, x => x.Count));
     }
 
-    private async Task<IActionResult> ProcessExcel(
-        int companyId,
-        byte[] content,
-        string fileName,
-        string mappingFile = "wbr_register_mapping.yaml")
-    {
-        _logger.LogDebug("ProcessExcel for {file} ({size} bytes)", fileName, content.Length);
-
-        var mappingPath = Path.Combine(AppContext.BaseDirectory, "mapping", mappingFile);
-        if (!System.IO.File.Exists(mappingPath))
-        {
-            _logger.LogError("Mapping file not found at {path}, ProcessExcel returning '500 Internal Server Error'", mappingPath);
-            return _500Mapping(mappingPath);
-        }
-
-        var mapping = RegisterMapping.Load(mappingPath);
-
-        System.Text.Encoding.RegisterProvider(System.Text.CodePagesEncodingProvider.Instance);
-        using var ms = new MemoryStream(content);
-        using var reader = ExcelReaderFactory.CreateReader(ms);
-        var dataSet = reader.AsDataSet();
-        if (dataSet.Tables.Count == 0 || dataSet.Tables[0].Rows.Count <= 1)
-        {
-            _logger.LogDebug("ProcessExcel returning '400 Bad Request' - Excel file is empty");
-            return _400EmptyRegister();
-        }
-
-        var table = dataSet.Tables[0];
-        var headerRow = table.Rows[0];
-        var columnMap = new Dictionary<int, string>();
-        for (int c = 0; c < table.Columns.Count; c++)
-        {
-            var header = headerRow[c]?.ToString() ?? string.Empty;
-            if (mapping.HeaderMappings.TryGetValue(header, out var prop))
-            {
-                columnMap[c] = prop;
-            }
-        }
-
-        var register = new Register { FileName = fileName, CompanyId = companyId };
-        _db.Registers.Add(register);
-        await _db.SaveChangesAsync();
-
-        Reference reference = new() { Id = register.Id };
-
-
-        var orders = new List<WbrOrder>();
-        for (int r = 1; r < table.Rows.Count; r++)
-        {
-            var row = table.Rows[r];
-            var order = new WbrOrder { RegisterId = register.Id, StatusId = 1, CheckStatusId = 1 };
-            foreach (var kv in columnMap)
-            {
-                var val = row[kv.Key]?.ToString();
-                var propInfo = typeof(WbrOrder).GetProperty(kv.Value);
-                if (propInfo != null)
-                {
-                    if (propInfo != null && val != null)
-                    {
-                        try
-                        {
-                            // Convert string value to appropriate property type
-                            object? convertedValue = ConvertValueToPropertyType(val, propInfo.PropertyType, propInfo.Name);
-                            propInfo.SetValue(order, convertedValue);
-                        }
-                        catch (Exception ex)
-                        {
-                            _logger.LogWarning(ex, "Failed to set value '{Value}' for property {Property}",
-                                val, propInfo.Name);
-                            // Continue with next property rather than failing the whole import
-                        }
-                    }
-                }
-            }
-            orders.Add(order);
-        }
-
-        _db.Orders.AddRange(orders);
-        await _db.SaveChangesAsync();
-
-        _logger.LogDebug("ProcessExcel imported {count} orders and is returning Reference with ID: {id}", orders.Count, reference.Id);
-        return CreatedAtAction(nameof(UploadRegister), new { id = reference.Id }, reference);
-        
-    }
-
-    private static readonly CultureInfo RussianCulture = new("ru-RU");
-
-    private object? ConvertValueToPropertyType(string? value, Type propertyType, string propertyName)
-    {
-        if (string.IsNullOrWhiteSpace(value))
-        {
-            // For nullable types or reference types, return null
-            if (propertyType.IsClass || Nullable.GetUnderlyingType(propertyType) != null)
-                return propertyType == typeof(string) ? string.Empty : null;
-
-            // For value types, this will throw an exception if we don't handle specific cases below
-        }
-
-        // Get the underlying type if it's nullable
-        Type targetType = Nullable.GetUnderlyingType(propertyType) ?? propertyType;
-
-        // Handle common types with specific conversion logic
-        if (targetType == typeof(int) || targetType == typeof(Int32))
-        {
-            return int.TryParse(value, NumberStyles.Integer, RussianCulture, out int result) ? result : default;
-        }
-        else if (targetType == typeof(decimal))
-        {
-            // Handle both comma and dot as decimal separators
-            string normalizedVal = value?.Replace('.', ',') ?? "0";
-            return decimal.TryParse(normalizedVal, NumberStyles.AllowDecimalPoint, RussianCulture, out decimal result) ? result : default;
-        }
-        else if (targetType == typeof(double))
-        {
-            string normalizedVal = value?.Replace('.', ',') ?? "0";
-            return double.TryParse(normalizedVal, NumberStyles.AllowDecimalPoint, RussianCulture, out double result) ? result : default;
-        }
-        else if (targetType == typeof(bool))
-        {
-            // Handle various representations of boolean values
-            if (string.IsNullOrWhiteSpace(value))
-                return default(bool);
-
-            string normalizedVal = value.ToLower(RussianCulture).Trim();
-            string[] trueValues = ["1", "yes", "true", "да"];
-
-            if (trueValues.Contains(normalizedVal, StringComparer.InvariantCultureIgnoreCase))
-                return true;
-            else 
-                return false;
-        }
-        else if (targetType == typeof(DateTime))
-        {
-            return DateTime.TryParse(value, RussianCulture, DateTimeStyles.None, out DateTime result) ? result : default;
-        }
-        else if (targetType == typeof(DateOnly))
-        {
-            // Handle DateOnly type - try parsing date portion only
-            if (DateOnly.TryParse(value, RussianCulture, DateTimeStyles.None, out DateOnly result))
-                return result;
-
-            return default(DateOnly);
-        }
-        else if (targetType == typeof(string))
-        {
-            // For string properties, return the value directly or an empty string for null/empty input
-            return value ?? string.Empty;
-        }
-
-        // For other types, try using the default conversion
-        try
-        {
-            return Convert.ChangeType(value, targetType, RussianCulture);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "Could not convert '{Value}' to type {Type} for property {Property}",
-                value, targetType.Name, propertyName);
-            return targetType.IsValueType ? Activator.CreateInstance(targetType) : null;
-        }
-    }
 }
